fix pointer computation

+++ src/datawizard/interfaces/bcsr_interface.c
@@ -119,9 +119,9 @@ void starpu_bcsr_data_register(starpu_data_handle_t *handleptr, unsigned home_no
 		STARPU_ASSERT_ACCESSIBLE(nzval);
 		STARPU_ASSERT_ACCESSIBLE(nzval + nnz*elemsize*r*c - 1);
 		STARPU_ASSERT_ACCESSIBLE(colind);
-		STARPU_ASSERT_ACCESSIBLE(colind + nnz*sizeof(uint32_t) - 1);
+		STARPU_ASSERT_ACCESSIBLE((uintptr_t) colind + nnz*sizeof(uint32_t) - 1);
 		STARPU_ASSERT_ACCESSIBLE(rowptr);
-		STARPU_ASSERT_ACCESSIBLE(rowptr + (nrow+1)*sizeof(uint32_t) - 1);
+		STARPU_ASSERT_ACCESSIBLE((uintptr_t) rowptr + (nrow+1)*sizeof(uint32_t) - 1);
 	}
 #endif
